requirements.txt
acme==1.1.0
appdirs==1.4.3
APScheduler==3.6.3
apturl==0.5.2
asgiref==3.6.0
astroid==2.4.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
bottle==0.12.19
bottle-websocket==0.2.9
Brlapi==0.7.0
bs4==0.0.1
certbot==0.40.0
certbot-nginx==0.40.0
certifi==2019.11.28
cffi==1.15.1
channels==3.0.5
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
ConfigArgParse==0.13.0
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.0
cupshelpers==1.0
daphne==3.0.2
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==4.2.1
django-crispy-forms==1.11.1
django-crontab==0.7.1
django-debug-toolbar==3.2
django-environ==0.4.5
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-xml==2.0.0
djangorestframework-yaml==2.0.0
drf-yasg==1.21.5
duplicity==0.8.12.0
Eel==0.14.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
geojson==2.5.0
gevent==21.1.1
gevent-websocket==0.10.1
greenlet==1.0.0
gunicorn==20.1.0
httpie==2.4.0
httplib2==0.14.0
hyperlink==21.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==21.3.0
inflection==0.5.1
isort==5.7.0
itypes==1.2.0
Jinja2==3.1.2
josepy==1.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.2
mccabe==0.6.1
mock==3.0.5
monotonic==1.5
more-itertools==4.2.0
mysqlclient==2.0.3
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
packaging==23.1
paramiko==2.6.0
parsedatetime==2.4
pbr==5.6.0
pexpect==4.6.0
phonenumbers==8.12.19
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.9.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
Pygments==2.9.0
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyodbc==4.0.30
pyOpenSSL==23.0.0
pyowm==3.1.1
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
pyTelegramBotAPI==3.8.2
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-telegram-bot==13.3
pytz==2023.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
SecretStorage==2.3.1
service-identity==21.1.0
simplejson==3.16.0
six==1.14.0
soupsieve==2.2
sqlparse==0.4.1
systemd-python==234
telebot==0.0.4
testresources==2.0.1
toml==0.10.2
tornado==6.1
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.3.0
tzlocal==2.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
whichcraft==0.6.1
whitenoise==5.3.0
wrapt==1.12.1
xkit==0.0.0
zipp==1.0.0
zope.component==4.3.0
zope.event==4.5.0
zope.hookable==5.0.0
zope.interface==5.2.0
